+++ src/app.ts
@@ -24,6 +24,7 @@ const config = {
     components: '*'
   }
 };
+
 try {
   const harmony = Harmony.load([BitCliExt, BitExt], config);
   harmony