+++ src/jsdoc/react/react-parser.ts
@@ -110,7 +110,7 @@ export default async function parse(data: string, filePath?: PathOsBased): Promi
       return formatted;
     }
   } catch (err) {
-    logger.debug(`failed parsing docs using docgen on path ${filePath} with error`, err);
+    logger.silly(`failed parsing docs using docgen on path ${filePath} with error`, err);
   }
   return undefined;
 }